Refactor sync-server out of index
